Move categorization_service to categorization package
Pure path rename: src/moneybin/services/categorization_service.py becomes
src/moneybin/services/categorization/__init__.py with no content changes.
All 28 import sites across MCP tools, CLI commands, services, scripts, and
tests updated from moneybin.services.categorization_service to
moneybin.services.categorization.

First step of splitting the 2,319-line service into a facade plus
collaborators (matcher, applier, assist) along the seams documented in
docs/specs/categorization-matching-mechanics.md, categorization-cold-start.md,
and categorization-auto-rules.md. No behavior change; 2058 tests pass.
