+++ src/moneybin/mcp/tools/merchants.py
@@ -15,7 +15,7 @@ from moneybin.database import get_database
 from moneybin.mcp._registration import register
 from moneybin.mcp.decorator import mcp_tool
 from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
-from moneybin.services.categorization_service import (
+from moneybin.services.categorization import (
     CategorizationService,
     validate_match_type,
 )
