+++ tests/moneybin/test_categorization_service_items.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from moneybin.services.categorization_service import (
+from moneybin.services.categorization import (
     CategorizationItem,
     CategorizationService,
     validate_items,
